productos/requirements.txt
Flask
python-dotenv
PyJWT